+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Next Release
+
+- **[NEW]** Add `BufferedLogger`, which captures logged messages in memory
+
 ## 1.0.0 (2018-03-06)
 
 - Initial release